LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/fetch/api/credentials/authentication-basic-expected.txt

http://localhost:8800/fetch/api/resources/authentication.py?realm=test - didReceiveAuthenticationChallenge - Simulating cancelled authentication sheet
http://localhost:8800/fetch/api/resources/authentication.py?realm=test - didReceiveAuthenticationChallenge - Simulating cancelled authentication sheet
http://localhost:8800/fetch/api/resources/authentication.py?realm=test - didReceiveAuthenticationChallenge - Simulating cancelled authentication sheet

FAIL User-added Authorization header with include mode assert_equals: HTTP status is 200 expected 200 but got 401
FAIL User-added Authorization header with same-origin mode assert_equals: HTTP status is 200 expected 200 but got 401
FAIL User-added Authorization header with omit mode assert_equals: HTTP status is 200 expected 200 but got 401
